container/app/src/requirements.txt
fastapi==0.111.0
fastapi-csrf-protect==0.3.3
starlette-csrf==3.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
alembic==1.13.1
pytest==8.2.0
injector==0.21.0
pytz==2024.1
ulid-py==1.1.0
PyJWT==2.8.0
